beginning major optimization branch
in the best practices section of the readme (or somewhere, anyway) i mention that if cycles aren't a problem you're better off efficiency-wise using shared_ptr due to the overhead of gc management.
however, i don't like that distinction - it makes it a bit tedious to think about and makes it harder to be efficient and use the library together.

my solution: keep all the current stuff for gc management and add duplicates of al the ptr<T> and make/adopt stuff that only uses reference counting.
this would make it significantly faster, wouldn't need to lock gc shared resource synchronizers, and would make graph traversals vastly faster because they wouldn't event need to be included in the graph.

this commit adds the first step of the process: a type called has_trivial_router (and its plural all_have_trivial_routers) that returns compile-time information on if type T has trivial router functions (i.e. ALL of its router functions are no-op).
if such a condition is met, we can selectively use the reference-count-only approach at compile-time for any type T.

added these intrinsics for all the pre-existing stdlib types and their equivalent wrappers.

the only potential issue is in the conversion system:
i.e. i'll need to figure out how the cast from ptr<T> to ptr<U> works if one is trivial and the other isn't.
i don't want to remove that feature, so i'll need to come up with some way to patch it.