extra und checks for GC::adopt()
realized that GC::adopt(T *obj) would be und if obj was not of type T (e.g. pointer to base) because it would reference the wrong router functions.
added this comment to the GC::adopt() comments.

added a preprocessor flag DRAGAZO_GARBAGE_COLLECT_EXTRA_UND_CHECKS that enables extra und checking at runtime.
this is currently only used for safeguarding the above GC::adopt() und case where T is a polymorphic type.
this is only done for polymorphic types because polymorphic types know their true type (though we can't access it) via operator typeid and we can thus verify its runtime type is exacly T.
non-polymorphic types have no such utility and thus cannot be checked - this is left as pure, unchecked undefined behavior (that is well-documented).

additionally, if C++17 is enabled, this check is wrapped in a constexpr if that is only compiled if T is a polymorphic type - thus (in C++17) this runtime check would not impact performance for non-polymorphic adoption operations (which are likely the most common type).

added this und usage of GC::adopt() to the undefined behavior section of the readme.

added assertion macros for throwing/not throwing any/specific exceptions.
this is used in a preprocessor-enabled test case in the test file for the new polymorphic GC::adopt() checking.