is_trivial optimization/feature
sad reacts.
a previous commit added the is_trivial system to routers in the hopes of being able to create separate systems for the gc and pure-ref-count systems so that the gc database would be as small and fast as possible.
unfortunately (as hinted in that commit's msg) the fact that a gc type can inherit from a non-gc type throws a wrench in that whole concept.
the gc type ptr would use gc logic and rooting/etc - when casting it to base (as cpp-gc provides) it would use reference count only.
perhaps even more perplexing: a gc ptr might be reassigned to a ref-count-only object, in which case it would need to unroot among other things - then it might be reassigned back to a gc object and need to re-root.

this is of course possible to do on the object level, but it would require the same kind of overhead as polymorphic calls and would be significantly-harder to write.
so i'm not planning on doing it any time soon, but it's a possibility.

on the bright side, i found a different use for it for the time being.
GC::route() and esp. GC::route_range() are now SFINAE-upgraded to be no-op for trivial types.
thus, if is_trivial is true the router functions are ignored entirely.
here's the best part: if is_trivial is true, you no longer even need to define the router functions.

this probably won't have much impact for GC::route() with optimizations enabled.
however, it could improve GC::route_range() because the compiler might not be able to elide the iteration loop otherwise (e.g. if ++iter or *iter has side-effects or makes non-trivial logic or refrences volatile data).

more importantly, however, by doing this we guarantee route() and route_range() will not be called for types marked with is_trivial.
thus GC::ptr<std::vector<T>> where T's router has is_trivial set to true will never be routed over - so it's totally safe to modify it however you want without mutexing.

this would likely be true regardless for vector because they typically use raw pointers for iterators.
however for something like std::deque it's possible that iterating and insering/deleting without mutexing the two could lead to und memory accesses e.g. if the array of subarrays deque approach is used and a new sub-array is inserted/removed.

thus this is primarily a safety feature + performance in multithreaded environments (i.e. most environments, due to how cpp-gc functions internally).