bug fix + code cleanup
silly yesterday me forgot all about iterator_traits and wrote my own (exposed) hack of it in the gc class.
it works for most iterators, but not for wrappers like std::vector<bool>::iterator which is undefined by the standard and apparently doesn't work with the system i used for getting iterator deref types.
my hack is replaced entirely with typename std::iterator_traits<Iter>::value_type, which is how it should have been in the first place.
this works for any standard-compliant iterator type, which is exactly what i'd intended.

also, since this is stdlib code, we obviously dont need test cases for it anymore.

only other change is a cleanup to the __gc_scoped_lock class for 2+ lockable targets (the non-trivial version):
old solution had copy-pasted code from a previous version of the project in which it was entirely inside the gc class and could have the unlocker outside of the lock class and yet still hidden by a private flag - thus it had residual redundant templating.
also it unlocked the mutexes in reverse order, which while probably not a big deal is kinda weird.

new solution's unlocking code is wrapped into a single function and exploits the fact that it can A) be a member function and B) doesn't directly source the Lockable parameter pack.
using these two facts, a shorter/sweeter function was written that's easier to use and unlocks them in order.