more optimizations
i've had my eye set on fixing the issue of recommending people use the stdlib container wrappers for use in gc objects but them being less efficient due to the mutexing.
fortunately, with our handy dandy little is_trivial feature (along with the last commit that made trivial routers always no-op and never iterate over the container) this is no longer an issue.

the wrapper types have been moved outside of the GC class and given names like __gc_etc - they have otherwise been unchanged.
inside the GC class, however, we use aliases to select the optimal choice between the stdlib container and the wrapper.
if the stdlib container's router is flagged as trivial it is used, which avoids the mutexing with no runtime performance loss, otherwise the wrapper is selected, which actually does the mutexing logic.

since i never advertised that the wrappers were threadsafe this is a perfectly acceptable change.

i don't expect this to be much of an improvement in terms of speed; it's mostly just so GC::etc and std::etc are more compatible in more cases (and now they might even be the same type).