improved some test file stuff
added macros for timer start/stop with automatic formatted messages and elapsed time stuff.
replaced the old timer usage with this - much more convenient.

noticed in the output there are some std::invalid_argument exceptions being thrown.
i put try blocks around all the thread init lambdas to make sure it wasn't any of them.
it wasn't any of my threads, so either it was a stdlib thread that it handled or hitting stop on the debugger causes an exception.
i'm moreso thinking it's the latter.

currently, whenever you want to do any gc actions (e.g. add new gc object to the obj list, repoint a ptr, construct/destruct a ptr, etc.) it has to lock the whole gc system's mutex.
thus if you have several threads making a lot of gc actions they'll all be blocking one another.
and worse, the problem gets worse the more threads are runnng (likely linear in the number of threads performing frequent gc operations, but maybe worse due to thread sleeping rather than spin waiting).

i've already looked into making each sub-system have a different mutex, but when one is used, it's extremely likely all are used so this would only hurt performance by increasing the number of locks we need to take (esp. anti-deadlocking algorithms for potential lock inversions).

however, i've thought of a new algorithm for the system that i think may be faster (usually O(1) in the number of threads, depending on usage).
the plan is to wrap all the gc database info (i.e. obj list and roots set, plus various control flags) into a single type and make each thread have a thread_local instance of it.
that way each thread has its own system to mutex on and thus (usually) won't interfere with one another.
of course, sharing gc objects among several threads would go back to the same problem, but only for THOSE objects and only for THOSE threads, rather than for EVERY object and for EVERY thread - that's the improvement.
in fact, in practice it's extremely uncommon to share resources across threads (relative to the number of resources kept uniquely by each thread), which is why i believe this would have a typical usage complexity of O(1) in the number of threads performing frequent gc operations.

the collector would be rewritten to lock all thread_local subsystems and aggregate all the info it needs to perform the operation across all threads, just as it does now but with the information more spread out.