code cleanup + minor safety patch
cleaned up a lot of the private code stuff.
there were a lot of helper functions/types scattered all over the place for various purposes (e.g. for tuple routing).
where possible, these were merged into one type, thus eliminating the externally-accessible helpers as well as frequent changes between public/private in the GC class.

made the aligned allocator system a bit cleaner:
still functions the same but now includes type aliases for the checked/unched variants for any alignment.
this eliminates the old system where we were directly using decltype sfinae logic.
additionally, the aliases take a variadic pack of alignment requirements and it automatically uses the highest of them.
this cleans up the allocator type access significantly.

fixed a missing safety case for routers: what if default impl is selected because T was ref qualified?
added std::remove_reference_t to both GC::route() as well as the static_assert safet net in the default router impl.
fortunately, this didn't catch anything in the test file, so it's probably pretty uncommon, but it's better to test it.
the only thing i could see causing this is routing to a forwarding reference, but that should (theoretically) never happen in a router function.