GC::alias() + global init order und fix
added GC::alias() for aliasing subcomponents of a gc object.
was previously only offered for the array for of ptr<T>::get(size_t) where it would alias an element of an array.
this has been renamed to ptr<T>::alias(size_t) and is mentioned to be a convenience form for the raw GC::alias() call.

c++ is a bit annoying in terms of global variable initialization order, which is undefined.
for instance, if a user declared a GC::ptr<T> at global scope it could potentially (depending on compiler) be initialized before the static variables used for controlling gc logic.
said ptr<T>'s constructor would thus reference the (undefined) static gc control variables and crash the program immediately before main() was even invoked.

of course, the fix for this is a well-known idiom - put the shared resources (in this case the previously-static) gc vontrol variables inside a function as a static local variable and return a reference to them.
there is now a class called GC::disjoint_module (previously the static class collection_synchronizer) which holds all this information in member variables rather than statics.
a factory accessor GC::disjoin_module::shared() defines a static local of this type and returns a reference to it, thus fixing the problem.

additionally, this serves another purpose for a potential update to come:
i've done some diagramming and proofing and have decided that, due to the nature of how the roots database works due to allowing shared resources, i can't do what one of the recent commits suggested where each thread gets its own submodule that limits mutex contention among several gc-capable threads.
the reason for this is thread A could add a mutable root to object X, but thread B calls unroot recursively on X.
the only way to do this would be to unroot from all roots submodule caches, but that would cripple performance.

of course, the only reason that this problem exists is because object X is shared by threads A and B, so either can modify it by adding/removing roots and any such change must be made to both threads' submodules.

thus, a new potential update it to make these submodules be defined as disjoint (hence the name choice for the new disjoint_module).
this means each thread would (potentially) get its own submodule, but they can also share.
threads that share the same disjoint module would be allowed to share objects.
threads with different disjoint modules would not be allowed to share objects.

so that's an optimization possibility that would actually work, but if i can't write in a way to ensure the undefined behavior case can't happen silently i won't do it.
on the other hand, if i can it would be an optional feature that the user could enable on a per-thread basis, and it would have its own set of functions for setting up sharing.