(abusive) wrapper deadlock fix
added a new ignore collection system.
each disjoint module now has an ignore collect count (the number of current active requests to ignore GC::collect()).
this is controlled by synchronized methods of the disjoint module class.

this is only exposed to the user (and is only used even in my impl code) as a sentry object called ignore_collect_sentry.

the ignore logic, although defined in terms of GC::collect() is actually performed by disjoint_module::collect().
this is to make sure it has as much exposure to the system as possible and that in the future i won't accidentally make a loophole around it.

i realized that if the ctor/dtor/asgn of type T in a gc wrapper e.g. std::vector<T> has calls to GC::collect() in the ctor/dtor/asgn it can deadlock.
this is caused by the mutating methods of the wrapper locking the router mutex - then when the call to collect is made (with the router mutex still locked) the collector routes to the object and deadlocks on the router mutex which is still locked.

added two prevention mechanisms, both of which are toggleable via preprocessor directives at the top of the header file.
one fix (the default option) is to use a recursive mutex instead of a normal mutex inside the wrappers.
this fixes the problem of having calls to GC::collect() directly from the same thread in the ctor/dtor/asgn.

the other fix (implemented, but toggled off by preprocessor) is to create an ignore_collect_sentry before calling the arbitrary code in the wrapper methods that need to lock the router mutex.
this is better than the other solution in that it catches cases like T::T() invoking GC::collect() asynchronously from another thread and then blocking until it's done (another one of my new test cases).
however, you'd have to go out of your way to screw that up, so it's a bit overkill - the other solution will cover most every case.
also, it could be a bit slower due to the extra disjoin module mutex locking.
and this doesn't even mention the headache of this causing spurious GC::collect() failures if there just happens to be an ignore collect action that was started by another thread mutating a wrapper that happens to exist in your disjoint module - it would make calls to GC::collect() potentially-meaningless.
with this disabled, the only way for GC::collect() to fail spuriously is if you, the user, opted to enable the wrapper setting or if you yourself created the ignore_collect_sentry.

added a disavowal of calling GC::collect() from ANY ctor/dtor/asgn function anywhere for any reason to the readme best practices section.