added disjunction stuff and tested performance
added the disjunction logic stuff - super hype:

the first time and gc utilities are used by a thread instantiates a thread_local disjunction handle (internally shared_ptr) whose default value is the primary disjunction (which is also created if it hasn't been already).
the primary disjunction is thus the default disjunction for a thread created by any mechanism, be it raw pthreads, some fancier os-stuff, or good old std::thread.
when the reference count for said disjunction hits zero the disjunction is deleted (hence the shared_ptr).

added a disjunction collection type that's implemented as a list of weak_ptr to disjunction modules.
the weak_ptr allows the disjunctions to be deleted whenver they need to without having to bog down resources until it can deal with it.
this class has (BACKGROUND-COLLECTOR-THREAD-ONLY) functions for collecting and culling.
these are only for the background thread because there's no reason anyone else should call them and it's more efficient + easier not to make it as thread-safe as a full-fledged disjunction module.
collection goes through the list and locks the weak_ptrs - if it's dead it deletes it from the list - otherwise it performs a collection on that disjunction module.
the culling mode only performs the dangling weak_ptr erasure step - called by the background collector when it would perform a full collect but the timed collect flag is disabled.

the only other function in this disjunction collection class allocates a new disjunction and returns a handle to it (shared_ptr).
this also inserts a weak_ptr version of it into the handle collection.
this function - unlike the other two - can be called from any thread at any time - and is non-blocking.
the non-blocking behavior is facilitated by an additional disjunction add cache, which is used instead when in collection mode from the background thread utility functions.
this cache follows the same logic as the various caches in the disjunction modules themselves - i.e. should be empty when not collecting - must be used while collecting.

added a cpp-gc "wrapper" for std::thread called GC::thread.
this functions much like the other wrappers but actually adds/removes behavior from the std interface that it wraps.
everything is the same except the constructors.
std::thread constructor takes a function + args.
GC::thread takes a disjunction mode + function + args.
these modes are: primary_disjunction_t, inherit_disjunction_t, and new_disjunction_t.

primary_disjunction acts as std::thread (and any other type of threading) - passing the (default) option is required to ensure you don't do it accidentally.

inherit_disjunction inherits the disjunction handle from the thread that is constructing the GC::thread object - this is facilitated by defining a new perfectly-forwarding lambda that additionally takes a shared_ptr handle to the disjunction module to use - it repoints the thread_local handle to the parent disjunction and then calls the original function.

new_disjunction functions much like inherit_disjunction except that it doesn't take a disjunction handle parameter and instead creates and repoints the thread_local handle to a new disjunction created by the (static) disjunction collection object.

a while back the interference test was added to the test file specifically to test the performance of this feature when it was eventually added.
here are the statistics regarding this:
before disjunctions:  ~9100             ms
primary_disjunction: N(9189.3,  91.778) ms
inherit_disjunction: N(9112.5, 180.387) ms
new_disjunction:     N(2937.3, 222.936) ms

these statistics were taking with a sample size of 10 in release mode.

primary/inherit should be much the same because they're coming from a thread that was already in the primary disjunction.
also, they're approximately the same as before this commit, which we should also expect since they're all still in the same disjunction.
but behold the holy grail: new_disjunction is substantially-faster than any of the others. - which is exactly what we'd expect if it was working :D !!

i even added a thread-safe printing function and had the threads print out a message when they were done to make sure something weird wasn't happening.
in all modes all threads successfully finished their work and printed the message.

so it works, super hype.

now all i have to do is address destructors not being called on program exit and this bad boi will be good to go.
this wasn't addressed earlier because without the ability to create new disjunctions testing if it was correctly deleting everything on program end would've required the program to actually end.
now i can just simulate it ending by creating a new disjunction thread and ending the thread normally - much easier.