GC::thread testing (dangling reference tests)
i was a bit concerned that the lambda used inside the fancy constructors of GC::thread were taking their arguments with forwarding references.
specificially, i was concerned that the references would be copied, control would resume in the thread that created the new GC::thread, and then the objects could potentially be destroyed before the GC logic inside the lambda was finished and the references were forwarded to the function object.

so i did some testing and determined that that can never happen.
in fact, it turns out before the function passed to the std:thread constructor is even called the args are forwarded to decay functions that create by-value copies on the new thread's stack.
this all happens before control returns to the thread that called the constructor.
and since they're passed through decay functions and stored as copies on the new stack there's no chance of accessing undefined memory.

once on the new thread's stack, the args are sent to the function object as xvalues.
thus, i changed the auto&&... pack in the lambda to std::decay_t<Args>&&... because we're guaranteed they'll be xvalues (and non-const due to the decay copy).
they're then passed to the function as a std::move(args)... pack expansion.
the same logic is also applied to the function object.

also, the shared_ptr disjunction handle is now an xvalue ref param as well, which avoids an unnecessary atomic inc/dec.
this is safe because we're guaranteed the decay copy already created a copy of the ptr elsewhere and is just being moved as an xvalue into the function (i.e. it still exists out there in the new thread's stack and thus won't let the ref count drop to zero).

after object intrinsics were performed (# move/cpy/etc), i've determined the wrapper constructors are safe and perform just as many move/copy operations as the std::thread's constructor.
so this type is good to go.