linux touchups + more GC::thread safety checks
moved back over to linux to make sure everything's working fine.
there were some compile errors that were introduced from the more-lenient visual studio compiler.
those are now fixed - much of that was just fixing the "specialization outside of namespace scope" error that i do a lot.

performed some more security scrutiny of the GC::thread wrapper in the linux distribution - everything seems to be fine.
one oddity is that the linux build performs one extra move construction than the Windows version (which also occurs for std::thread).
however, i put logging messages in in the GC::thread constructor's lambda and found that these extra move constructors vere before the lambda body.
additionally, the lambda parameters are all rvalue references, which means invoking the lambda wouldn't trigger any constructors.
therefore i've come to the conclusion that it's just an oddity of how the linux system handles the process of moving the data around for the thread, and just happens to be less efficient than the windows version.
this isn't really a problem, and even if it were i can't do anything about it with only standard C++ at my disposal.

at any rate, it appears that the efficiency for the GC::thread wrapper - even in the linux system - is equal to std::thread and is safe.
i added a test case for passing a vector to a function in another thread and writing it to stderr and could not produce an error.
i was also unable to find any cause for doubting the safety of the GC::thread constructors.

so GC::thread seems to be perfectly safe.