weird massive memory usage bug fix
apparently in the changes from the last commit the memory usage went from 1 MB (peak 2MB) to 10-12 MB, whch it never drops from.
after some investivation it was apparently caused by the thread_local test object (was just a test to make sure thread_local things were created and destroyed when the actual threads were).
i'm still not really sure how it could've lead to that kind of memory usage, but after removing only the printing line from the synchronized printer function is was using (leaving in the mutex) the problem went away.
i even tried adding in a thread sleep to see if a timeing thing along with the mutexing was what was causing the problem, but it appears that even that's not the case...

after doing some digging it seems some compilers used to not even allow thread_local objects with non-trivial ctor/dtor, so it's possible it has something to do with that.