added first round of disjunction error tests
at this point disjunctions are so volatile (due to immediately killing the process if you get them wrong) i'd almost suggest people don't even use them and just leave everything in the primary (default) disjunction.
however, they're really good for performance-centric code, so if i can just make them safer to use it'll be great for users.

this commit adds GC::disjunction_error : std::runtime_error.
this exception type is thrown by anything that detects a disjunction violation.
additionally, if an exception of this type is thrown the violating operation is not performed.
this means the cases checked by this system won't crash the whole program on violation - they'll just emit a normal exception.

currently this is only used for repointing a GC::ptr (not including swap, which i'll add in with the next batch).
added test cases for this in all current GC::thread disjunction modes.

as a consequence, GC::disjoint_module had to be rewritten (just the func sigs) in terms of smart_handle instead of raw info*.
this is because the repoint check requires knowing what disjunction the smart_handle belongs to (i.e. if smart_handle's and info's disjunctions are different, triggers an exception).

this also means smart_handle has an extra pointer inside it when the disjunction checks are enabled.
consequently, a GC::ptr is now 3 machine words instead of 2 and no longer fits in an xmm register in 64-bit mode.
this means hardware transfers of GC::ptr will be about twice as slow (unless it for some reason uses a ymm register, but i doubt it).
however, that's in all likelyhood negligible on most systems (it is on mine at least).