more disjunction safety checks
added disjunction safety checks for swap and alias (copy) construction.
like the other disjunction safety checks, if this throws the violating action is not performed and undefined behavior is not invoked.

at first i was unsure if the alias (copy) constructor should enforce disjunction rules.
my counter-example at the time was "what if i need to pass a GC::ptr to another thread in another disjunction".
however, i now consider the alias (copy) constructor of GC::ptr (internally of GC::smart_handle) to be equivalent to the null (default) constructor plus a repoint action.
thus it is illegal and must be checked.
if you ever encounter the need to pass a GC::ptr to a new thread in a different disjunction you should pass by reference - or if lifetime is a concern, pass a GC::unique_ptr<GC::ptr> or GC::shared_ptr<GC::ptr> (or their std equivalents - wouldn't matter since they're not part of a gc allocation object).