local disjunction repoint bug fix
fixed a bug where i wasn't accounting for the fact that the local disjunction can be repointed after an object relying on it has already been constructed.
e.g. the test case i added that broke this was passing a GC::ptr by value to a GC::thread constructor in GC::new_disjunction mode.
the GC::ptr was constructed in the primary disjunction (added as a root there), the disjunction was repointed to a new one, the destructor was called which performs destruction logic in the local (new) module, thus leaving the root in the primary disjunction - even worse it now points to und memory, which was causing a crash during the subsequent calls to collect().

this is fixed by always referencing the disjunction pointers in the GC::smart_handle and GC::info types instead of disjoint_module::local().
of course this also means the pointers can no longer disabled by preprocessor options.
this means there's a minor memory usage hit in the optimized release build, but at least it doesn't crash without warning if the user does this (which the user shouldn't, by the way).

passing GC::ptr by value to thread constructors other than GC::inherit_disjunction may be added to the stuff-not-to-do section of the readme in the future.
first i'll see if there's an actual problem with doing it now that the und access error is fixed.
at the very least it might lead to unexpected behavior, as i don't believe the standard specifies which thread constructs the objects into the function call.

like the last commit mentioned, the safest way to pass a GC::ptr to a thread ctor is inside a GC::unique_ptr, GC::shared_ptr, or a global variable.
if lifetime's not an issue, GC::ptr<T>* would also suffice.