more thread disjunction tests
apparently any thread not handling an exception triggers an abort() - i assumed it was like C# threads where that thread stops but nothing else is affected.
removed the therefore-useless assert_throw/nothrow macros from the thread disjunction tests.

added value to value GC::ptr passes between threads.
only inherit disjunction mode should work - the others should all throw when the by-value function arg is constructed.
the other test cases (the ones that throw on thread construction) can't be added (commented out) because the uncaught exception would trigger an abort().

this commit tests all the new disjunction tests in the linux build.
i was suspicious the double move ctor the linux build was doing on thread creation would be a problem - this is fortunately not the case.
actually this is even stranger, as it implies both move ctors are performed by the old thread (i.e. the new thread isn't doing the second one).