new post-destruction gc module access test
added a test case that performs a GC::collect() after the disjunction handle has been destroyed (during disjoint module dtor due to shared_ptr ref count drop to zero).
disturbingly, this doesn't seem to cause an error and in fact the object is never destroyed at all.
i need to do more testing to find out why it isn't being destroyed (i.e. the dtor has "assert(false)" in it and that's not triggering).

once this is found and fixed i can get around to actually fixing the problem the test is supposed to address, which i already have an algorithm thought up for.

removed some references to disjoint_module::local() that are a bit risky depending on when they're called and on what objects, and when and what those objects happen to be.
this is fixed by using a pre-cached disjunction pointer from a smart_handle or an info object.
these are the cases where using disjoint_module::local() directly might cause an issue because the object/handle might have a conflicting module it belongs to.

in particular, mark_sweep() is potentially dangerous if disjunction violations occur due to using local() directly - this is changed to be a member function of GC::info and now uses the GC::info::disjunction value for performing the unrooting logic (as does the old unrooting router helper function).
these are a bit safer and faster due to not needing to enter the threadsafe local static ctor area.