fix local handle access after dtor
fixed a rather annoying bug where the call to collect() after a disjunction ref count hits zero was accessing the local handle after it was destoyed.
this was happening because local_handles are destroyed before the static (primary) handle.
this means the call to collect() on module ref count dec for this specific disjunction was causing und mem access.

this is now fixed by having a static detour pointer for the local accessors to use.
core handle features are now called primary_handle() and local_handle().
these should only be used if the actual handle is needed - e.g. for creating a new disjunction.
otherwise you should use the primary() and local() functions that return a raw pointer to the disjunction, accounting for local detours from the primary disjunction handle static dtor.
these return raw pointers because there aren't lifetime concerns when using prvalues as ret types.