code collapsing
collapsed a few similar code pieces:
BACKGROUND_COLLECTOR_ONLY__collect/cull are now 1 function that take a bool arg to switch between collect/cull - i can't believe i didn't do it that way the first time...
the background collector thread function is now gone - replaced with a lambda inside the start_timed_collect() function.

start_timed_collect() no longer performs a synamic allocation for the thread object.
this was originally done so that the destructor wouldn't be called on it (which would crash if the thread was still running).
new code has it created during a local static ctor in start_timed_collect(), where it is created as a temporary and then immediately detached.

what alerted me to the timed collect function was that it was showing up in the thread logging created from the last commit (in linux).
it only appeared when run on windows - perhaps the linux version's compiler is smart enough to elide it for this trivial example.

added a safety check for accessing local_handle() during static dtor time (when thread locals have already been destroyed).