fixed a background collector und access
fixed a bug where a thread (including the main thread) terminating during a background collection on its disjunction would trigger the assertions in the disjoint module dtor and kill the program.
this is in part caused by the background collector not using the iterated disjunction handle as its local handle.
this is also caused by no-op collections at the end of a shared_ptr handle's lifetime.
this is also caused by the last ref counted disjunction handle alive not being a local/primary handle.
also caused by several other things that are bad.

all of this is now solved - yay.
2 new types shared_disjoint_handle and weak_disjoint_handle have been added.
they're quite similar to std::shared_ptr and std::weak_ptr but have extra logic built into them for gc-disjunction specific stuff.
for instance, asgn/dtor trigger the end-of-thread cleanup logic.
this no longer has to be handled by the primary/local handle dtors.
in fact the problems above were caused by this same logic not being applied for repointing and also not being applied for all handles (was only on the primary/local handles, and only in the dtors).