C++17 port + bug fixes + touchups
changed over to all native C++17 stuff.
the home-brew C++17 stuff is now removed and also not aliased under the GC class.
removed all the options pertaining to C++ versions - all are C++17 now with no other options.

removed a few stupid things that i've sense disabled anyway because they're easy to abuse, dangerous, and promote bad programming styles: the gc ignore system.
the ignore system uses sentry objects that prevent collection during their lifetime.
however there are some problems: their construction doesn't synchronize with the end of a collection, so you're not guaranteed a collection isn't already in progress.
another is that it allows users to inadvertantly add loopholes into the cpp-gc guarantees pertaining to object lifetime esp. on explicit collection invocation.

the option for having ignore sentries in wrappers is now removed, and the sentries themselves will also likely be removed in the near future (they're currently still accessible).
on the other hand, i could count it as a feature (albeit a dangerous one) and leave it in with a massive warning about the various problems it could introduce.

fixed a bug where the default template for GC::wrapper_traits for a non-specialized cv family (e.g. int, float) was causing compile errors due to self-referential typedefs and template aliasing.
this was caused by the typedefs/aliases using std::conditional_t to select their cv qualification branches, one of which was itself in the case of a non-cv qualified and non-specialized type (hence the self-referential problem).
this is fixed by having the cv qualification branching done through sfinae indirection, with an added type indirection for preserving cv and ref qualifiers.

added compile time test cases for all of these new problems.